# codex/bundles/process-module-architecture-v3/proof/SB34-post-sb33-runtime-reliability-performance-and-e2e-review/swagger-process-routes.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CanDoItAll.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5032/"
    }
  ],
  "paths": {
...
    "/api/access/tokens": {
      "post": {
        "tags": [
          "API"
        ],
        "operationId": "IssueApiToken",
        "requestBody": {
          "content": {
            "application/json": "@{schema=}"
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}